+++ R/gbm.loop.R
@@ -22,13 +22,19 @@
 #' @param tc See gbm.auto help.
 #' @param lr See gbm.auto help.
 #' @param bf See gbm.auto help.
+#' @param n.trees See gbm.auto help.
 #' @param ZI See gbm.auto help. Choose one.
+#' @param fam1 See gbm.auto help. Choose one.
+#' @param fam2 See gbm.auto help. Choose one.
 #' @param simp See gbm.auto help.
 #' @param gridslat See gbm.auto help.
 #' @param gridslon See gbm.auto help.
+#' @param multiplot See gbm.auto help. Default False
 #' @param cols See gbm.auto help.
 #' @param linesfiles See gbm.auto help; TRUE or linesfiles calculations fail.
+#' @param smooth See gbm.auto help.
 #' @param savegbm See gbm.auto help.
+#' @param loadgbm See gbm.auto help.
 #' @param varint See gbm.auto help.
 #' @param map See gbm.auto help.
 #' @param shape See gbm.auto help.
@@ -37,6 +43,8 @@
 #' @param alerts See gbm.auto help; default FALSE as frequent use can crash
 #' RStudio.
 #' @param pngtype See gbm.auto help. Choose one.
+#' @param gaus See gbm.auto help.
+#' @param MLEvaluate See gbm.auto help.
 #' @param runautos Run gbm.autos, default TRUE, turn off to only collate
 #' numbered-folder results.
 #' @param Min.Inf Dummy param for package testing for CRAN, ignore.
@@ -93,18 +101,32 @@ gbm.loop <- function(loops = 10, # the number of loops required, integer
                      # lr = list(c(0.01,0.02),0.0001) or list(0.01,c(0.001, 0.0005))
                      bf = 0.5,             # permutations of bag fraction allowed, can be single
                      # number, vector or list, per tc and lr
+                     n.trees = 50,         # from gbm.step, number of initial trees to fit. Can be
+                     # single or list but not vector i.e. list(fam1, fam2)
                      ZI = "CHECK", # Are data zero-inflated? "CHECK"/FALSE/TRUE.
                      # Choose one.
                      # TRUE: delta BRT, log-normalised Gaus, reverse log-norm and bias corrected.
                      # FALSE: do Gaussian only, no log-normalisation.
                      # CHECK: Tests data for you. Default is TRUE.
+                     fam1 = c("bernoulli", "binomial", "poisson", "laplace", "gaussian"),
+                     # probability distribution family for 1st part of delta process, defaults to
+                     # "bernoulli",
+                     fam2 = c("gaussian", "bernoulli", "binomial", "poisson", "laplace"),
+                     # probability distribution family for 2nd part of delta process, defaults to
+                     # "gaussian",
                      simp = TRUE,          # try simplfying best BRTs?
                      gridslat = 2,         # column number for latitude in 'grids'
                      gridslon = 1,         # column number for longitude in 'grids'
+                     multiplot = FALSE,     # create matrix plot of all line files? Default false
+                     # turn off if large number of expvars causes an error due to margin size problems.
                      cols = grey.colors(1,1,1), # barplot colour vector. Assignment in order of
                      # explanatory variables. Default 1*white: white bars black borders. '1*' repeats
                      linesfiles = TRUE,   # save individual line plots' data as csv's?
+                     smooth = FALSE,       # apply a smoother to the line plots? Default FALSE
                      savegbm = FALSE,       # save gbm objects and make available in environment after running? Open with load("Bin_Best_Model")
+                     loadgbm = NULL,       # relative or absolute location of folder containing
+                     # Bin_Best_Model and Gaus_Best_Model. If set will skip BRT calculations and do
+                     # predicted maps and CSVs. Default NULL, character vector, "./" for working directory
                      varint = FALSE,        # calculate variable interactions? Default:TRUE, FALSE
                      # for error "contrasts can be applied only to factors with 2 or more levels"
                      map = TRUE,           # save abundance map png files?
@@ -114,9 +136,16 @@ gbm.loop <- function(loops = 10, # the number of loops required, integer
                      alerts = FALSE,        # play sounds to mark progress steps
                      pngtype = c("cairo-png", "quartz", "Xlib"), # file-type for png files,
                      # alternatively try "quartz" on Mac. Choose one.
+                     gaus = TRUE,          # do Gaussian runs as well as Bin? Default TRUE.
+                     MLEvaluate = TRUE,    # do machine learning evaluation metrics & plots? Default TRUE
+                     # brv = NULL, # addresses devtools::check's no visible binding for global variable https://cran.r-project.org/web/packages/data.table/vignettes/datatable-importing.html#globals
+                     # grv = NULL, # addresses devtools::check's no visible binding for global variable https://cran.r-project.org/web/packages/data.table/vignettes/datatable-importing.html#globals
+                     # Bin_Preds = NULL, # addresses devtools::check's no visible binding for global variable https://cran.r-project.org/web/packages/data.table/vignettes/datatable-importing.html#globals
+                     # Gaus_Preds = NULL, # addresses devtools::check's no visible binding for global variable https://cran.r-project.org/web/packages/data.table/vignettes/datatable-importing.html#globals
                      runautos = TRUE,      # run gbm.autos, default TRUE, turn off to only collate numbered-folder results
                      Min.Inf = NULL, # addresses devtools::check's no visible binding for global variable https://cran.r-project.org/web/packages/data.table/vignettes/datatable-importing.html#globals
                      ...) {
+
   # Generalised Boosting Model / Boosted Regression Tree process chain automater
   # Simon Dedman, 2012-8 simondedman@gmail.com github.com/SimonDedman/gbm.auto
 
@@ -126,9 +155,6 @@ gbm.loop <- function(loops = 10, # the number of loops required, integer
   # But based on what though? Just do a line of x:loop# vs y: minmin/maxmax/avav/avvar?
   # when change in variance from 1:2 to 1:3 to 1:n drops below a percentage threshold?
   # Fix csvs colnames, see https://github.com/SimonDedman/gbm.auto/issues/37
-  # for factorial variables, need to change from lines to bars
-  # Runautos doesn't work: binbars.df & gausbars.df are created & incrementally grown within the autos loop then accessed afterwards.
-  # Would need to do this separately somehow, possibly a separate loop to pull these data from a source file csv?
 
   # utils::globalVariables("Min.Inf") # addresses devtools::check's no visible binding for global variable https://cran.r-project.org/web/packages/data.table/vignettes/datatable-importing.html#globals
 
@@ -145,15 +171,10 @@ gbm.loop <- function(loops = 10, # the number of loops required, integer
     beep(9)
     graphics.off()})  # give warning noise if it fails
 
+  fam1 <- match.arg(fam1) # populate object from function argument in proper way
+  fam2 <- match.arg(fam2)
   pngtype <- match.arg(pngtype)
 
-  binbars.df <- data.frame(var = rep(NA, length(expvar)),
-                           rel.inf = rep(NA, length(expvar)))
-  gausbars.df <- binbars.df # blank dataframes for bin & gaus bars data
-  report.df <- data.frame(BinCV = rep(NA, length(loops)),
-                          AUC = rep(NA, length(loops)),
-                          GausCV = rep(NA, length(loops)))
-  if (calcpreds) var.df <- grids[,c(gridslon, gridslat)] # create df with just lat & longs
   if (runautos) { # run gbm.autos unless turned off
     for (i in 1:loops) { # loop through all gbm.autos
       dir.create(paste0("./", i)) # create i'th folder
@@ -165,13 +186,20 @@ gbm.loop <- function(loops = 10, # the number of loops required, integer
                tc = tc,
                lr = lr,
                bf = bf,
+               n.trees = n.trees,
                ZI = ZI,
+               fam1 = fam1,
+               fam2 = fam2,
                simp = simp,
                gridslat = gridslat,
                gridslon = gridslon,
+               multiplot = multiplot,
                cols = cols,
                linesfiles = linesfiles,
+               smooth = smooth,
+               savedir = "./", # current directory created & moved to above
                savegbm = savegbm,
+               loadgbm = loadgbm,
                varint = varint,
                map = map,
                shape = shape,
@@ -179,28 +207,46 @@ gbm.loop <- function(loops = 10, # the number of loops required, integer
                BnW = BnW,
                alerts = alerts,
                pngtype = pngtype,
+               gaus = gaus,
+               MLEvaluate = MLEvaluate,
                ...) # accept other gbm.auto values than these basics
-      setwd(paste0("./", colnames(samples[resvar]))) # set wd to species named subfolder
+      setwd("../") # move back up to root folder
+      if (cleanup) unlink(i, recursive = TRUE)
+      print(paste0("XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX      Gbm.auto loop ", i, " complete        XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX"))
+    } # close i loop & go to the next i
+  } # close runautos if optional
 
-      if (file.exists("Binary BRT Variable contributions.csv")) {
-        binbarstmp <- read.csv("Binary BRT Variable contributions.csv") # temp container for bin bars
-        if (i == 1) {binbars.df <- binbarstmp} else {# csv file to df unless df exists
-          binbars.df <- rbind(binbars.df, binbarstmp)} # if so add to bottom of existing
-        bin = TRUE} else bin = FALSE
+  binbars.df <- data.frame(var = rep(NA, length(expvar)),
+                           rel.inf = rep(NA, length(expvar)))
+  gausbars.df <- binbars.df # blank dataframes for bin & gaus bars data
+  report.df <- data.frame(BinCV = rep(NA, length(loops)),
+                          AUC = rep(NA, length(loops)),
+                          GausCV = rep(NA, length(loops)))
+  if (calcpreds) var.df <- grids[,c(gridslon, gridslat)] # create df with just lat & longs
+
+  for (i in 1:loops) { # loop through all gbm.autos
+    setwd(paste0("./", i, "/", colnames(samples[resvar]))) # set wd to loopnumber and species named subfolder
+
+    if (file.exists("Binary BRT Variable contributions.csv")) {
+      binbarstmp <- read.csv("Binary BRT Variable contributions.csv") # temp container for bin bars
+      if (i == 1) {binbars.df <- binbarstmp} else {# csv file to df unless df exists
+        binbars.df <- rbind(binbars.df, binbarstmp)} # if so add to bottom of existing
+      bin = TRUE} else bin = FALSE
 
       # loop thru variables name linesfiles e.g. Bin_Best_line_[varname].csv
       # adding i'th loop's values as new column
       if (bin) for (j in colnames(samples[expvar])) {
-        if (!file.exists(paste0("Bin_Best_line_", j, ".csv"))) {tmp <- data.frame(x = rep(NA,100), y = rep(NA,100))}
-        #if file not created because simp, populate with 0s
+        if (class(samples[,j]) == "factor") nreps = length(levels(samples[,j])) else nreps = 100
+        # rep 100 is fine for numeric variables but needs to match n(levels) of factorial variables.
+        if (!file.exists(paste0("Bin_Best_line_", j, ".csv"))) {tmp <- data.frame(x = rep(NA, nreps), y = rep(NA, nreps))}
+        #if file not created because simp, populate with NAs
         if (file.exists(paste0("Bin_Best_line_", j, ".csv"))) {tmp <- read.csv(paste0("Bin_Best_line_", j, ".csv"))}
         #else use values
 
         colnames(tmp)[2] <- paste0("Loop",i)
         if (i == 1) {assign(paste0("binline_", j), tmp)
         } else {
-          assign(paste0("binline_", j), cbind(get(paste0("binline_", j)),
-                                              tmp[,2]))
+          assign(paste0("binline_", j), cbind(get(paste0("binline_", j)), tmp[,2]))
           if (is.na(get(paste0("binline_", j))[1,1])) { #if the first cell is NA (all 1st col, x, is na)
             assign(paste0("binline_", j), #rebuild same obj as df
                    data.frame(x = tmp[,1], #start with this loop's x values, hopefully not NA also
@@ -213,93 +259,84 @@ gbm.loop <- function(loops = 10, # the number of loops required, integer
           gausbars.df <- rbind(gausbars.df, gausbarstmp)}
         gaus = TRUE} else gaus = FALSE
 
-      if (gaus) for (k in colnames(samples[expvar])) {
-        if (!file.exists(paste0("Gaus_Best_line_", k, ".csv"))) {tmp <- data.frame(x = rep(NA,100), y = rep(NA,100))}
-        #if the first loop is simplified then the first col of gausline will be NAs which should be the X for the linefiles
-        #else use existing csv file, 2 columns
-        if (file.exists(paste0("Gaus_Best_line_", k, ".csv"))) {tmp <- read.csv(paste0("Gaus_Best_line_", k, ".csv"))}
-        colnames(tmp)[2] <- paste0("Loop",i)
-        if (i == 1) {assign(paste0("gausline_", k), tmp)
-        } else {
-          assign(paste0("gausline_", k), cbind(get(paste0("gausline_", k)),
-                                               tmp[,2]))
-          if (is.na(get(paste0("gausline_", k))[1,1])) { #if the first cell is NA (all 1st col, x, is na)
-            assign(paste0("gausline_", k), #rebuild same obj as df
-                   data.frame(x = tmp[,1], #start with this loop's x values, hopefully not NA also
-                              get(paste0("gausline_", k))[,2:(i + 1)]))} #then add the remainder of the existing obj cols
-          #column cbound but not named. Can name as string "col name" = 1:10, or
-          #objectname ColName = 1:10 but not formulaicly paste0("Col","Name") = 1:10
-          #or anything evaluated e.g. colnames(tmp)[2] = tmp[,2]
-          #colnames(paste0("gausline_", k))[i + 1] <- paste0("loop", i) #rename last column (loop# + 1)
-        }}
-      if (!file.exists("Abundance_Preds_only.csv")) calcpreds = FALSE
-      if (calcpreds) {predtmp <- read.csv("Abundance_Preds_only.csv") # temp container for latest preds
-      var.df <- cbind(var.df, predtmp[,3]) # cbind preds to existing lat/longs or other preds
-      colnames(var.df)[2 + i] <- paste0("Loop_", i)} # label newly added preds column
+        if (gaus) for (k in colnames(samples[expvar])) {
+          if (class(samples[,j]) == "factor") nreps = length(levels(samples[,j])) else nreps = 100
+          if (!file.exists(paste0("Gaus_Best_line_", k, ".csv"))) {tmp <- data.frame(x = rep(NA, nreps), y = rep(NA, nreps))}
+          #if the first loop is simplified then the first col of gausline will be NAs which should be the X for the linefiles
+          #else use existing csv file, 2 columns
+          if (file.exists(paste0("Gaus_Best_line_", k, ".csv"))) {tmp <- read.csv(paste0("Gaus_Best_line_", k, ".csv"))}
+          colnames(tmp)[2] <- paste0("Loop",i)
+          if (i == 1) {assign(paste0("gausline_", k), tmp)
+          } else {
+            assign(paste0("gausline_", k), cbind(get(paste0("gausline_", k)), tmp[,2]))
+            if (is.na(get(paste0("gausline_", k))[1,1])) { #if the first cell is NA (all 1st col, x, is na)
+              assign(paste0("gausline_", k), #rebuild same obj as df
+                     data.frame(x = tmp[,1], #start with this loop's x values, hopefully not NA also
+                                get(paste0("gausline_", k))[,2:(i + 1)]))} #then add the remainder of the existing obj cols
+            #column cbound but not named. Can name as string "col name" = 1:10, or
+            #objectname ColName = 1:10 but not formulaicly paste0("Col","Name") = 1:10
+            #or anything evaluated e.g. colnames(tmp)[2] = tmp[,2]
+            #colnames(paste0("gausline_", k))[i + 1] <- paste0("loop", i) #rename last column (loop# + 1)
+          }}
+        if (!file.exists("Abundance_Preds_only.csv")) calcpreds = FALSE
+        if (calcpreds) {
+          predtmp <- read.csv("Abundance_Preds_only.csv") # temp container for latest preds
+          var.df <- cbind(var.df, predtmp[,3]) # cbind preds to existing lat/longs or other preds
+          colnames(var.df)[2 + i] <- paste0("Loop_", i) # label newly added preds column
+        }
 
-      #Collect report CV & AUC scores
-      reporttmp <- read.csv("Report.csv") # temp container for bin bars
+        #Collect report CV & AUC scores
+        reporttmp <- read.csv("Report.csv") # temp container for bin bars
 
-      if ("Best.Binary.BRT" %in% colnames(reporttmp)) {
-        bincvtmp <- as.character(reporttmp$Best.Binary.BRT[2])
-        bincvspltmp <- strsplit(bincvtmp, "Model CV score: ")
-        bincvsplnumtmp <- as.numeric(bincvspltmp[[1]][2])
-        report.df[i,1] <- bincvsplnumtmp # copy BinCV score from this loop's report to allreport
+        if ("Best.Binary.BRT" %in% colnames(reporttmp)) {
+          # copy BinCV score from this loop's report to allreport
+          # cv.statistics$deviance.mean is already in gbm.auto report.csv: "CV Mean Deviance: ", but ISN'T in bin_best column!
+          # Therefore need to lookup best against other cols to find it.
+          binbestcomboname <- as.character(reporttmp$Best.Binary.BRT[1])
+          binbestcomboname <- strsplit(binbestcomboname, "Model combo: ")[[1]][2]
+          # replace lr1e- with lr1e. in the case of very small lr's, R will convert the colname to a dot and the lookup will fail
+          binbestcomboname <- gsub(pattern = "lr1e-", replacement = "lr1e.", x = binbestcomboname)
+          # if binbestcomboname is simplified, value ends in _Simp, but colname is "Simplified.Binary.BRT.stats"
+          # need to use length because if grep doesn't find the pattern it returns integer(0) which doesn't evaluate to logical FALSE for ==1 for some reason
+          if (length(grep(pattern = "_Simp", x = binbestcomboname)) == 1) binbestcomboname <- "Simplified.Binary.BRT.stats"
+          report.df[i, 1] <- as.numeric(strsplit(reporttmp[, which(colnames(reporttmp) %in% binbestcomboname)][3], "CV Mean Deviance: ")[[1]][2])
+          # copy AUC score from this loop's report to allreport
+          auctmp <- as.character(reporttmp$Best.Binary.BRT[3])
+          aucspltmp <- strsplit(auctmp, "Training data AUC score: ")
+          aucsplnumtmp <- as.numeric(aucspltmp[[1]][2])
+          report.df[i, 2] <- aucsplnumtmp
+          }
 
-        auctmp <- as.character(reporttmp$Best.Binary.BRT[3])
-        aucspltmp <- strsplit(auctmp, "Training data AUC score: ")
-        aucsplnumtmp <- as.numeric(aucspltmp[[1]][2])
-        report.df[i,2] <- aucsplnumtmp} # copy AUC score from this loop's report to allreport
+        if ("Best.Gaussian.BRT" %in% colnames(reporttmp)) {
+          # copy GausCV score from this loop's report to allreport
+          gausbestcomboname <- as.character(reporttmp$Best.Gaussian.BRT[1])
+          gausbestcomboname <- strsplit(gausbestcomboname, "Model combo: ")[[1]][2]
+          gausbestcomboname <- gsub(pattern = "lr1e-", replacement = "lr1e.", x = gausbestcomboname)
+          if (length(grep(pattern = "_Simp", x = gausbestcomboname)) == 1) gausbestcomboname <- "Simplified.Gaussian.BRT.stats"
+          report.df[i, 3] <- as.numeric(strsplit(reporttmp[, which(colnames(reporttmp) %in% gausbestcomboname)][3], "CV Mean Deviance: ")[[1]][2])
+        }
 
-      if ("Best.Gaussian.BRT" %in% colnames(reporttmp)) {
-        gauscvtmp <- as.character(reporttmp$Best.Gaussian.BRT[2])
-        gauscvspltmp <- strsplit(gauscvtmp, "Model CV score: ")
-        gauscvsplnumtmp <- as.numeric(gauscvspltmp[[1]][2])
-        report.df[i,3] <- gauscvsplnumtmp} # copy GausCV score from this loop's report to allreport
-
-      setwd("../../") # move back up to root folder
-      if (cleanup) unlink(i, recursive = TRUE)
-      print(paste0("XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX      Loop ",i," complete        XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX"))
-    } # close i loop & go to the next i
-  } # close runautos if optional
-
-  #runautos fix####
-  #move runautos binbars.df gausbars.df code to here, in separate loops
-  # binbars.df <- data.frame(var = rep(NA, length(expvar)),
-  #                          rel.inf = rep(NA, length(expvar)))
-  # gausbars.df <- binbars.df # blank dataframes for bin & gaus bars data
-  #
-  # if (file.exists("Binary BRT Variable contributions.csv")) {
-  #   binbarstmp <- read.csv("Binary BRT Variable contributions.csv") # temp container for bin bars
-  #   if (i == 1) {binbars.df <- binbarstmp} else {# csv file to df unless df exists
-  #     binbars.df <- rbind(binbars.df, binbarstmp)} # if so add to bottom of existing
-  #   bin = TRUE} else bin = FALSE
-  #
-  # if (file.exists("Gaussian BRT Variable contributions.csv")) {
-  #   gausbarstmp <- read.csv("Gaussian BRT Variable contributions.csv") # temp container for Gaus lines
-  #   if (i == 1) {gausbars.df <- gausbarstmp} else {
-  #     gausbars.df <- rbind(gausbars.df, gausbarstmp)}
-  #   gaus = TRUE} else gaus = FALSE
+        setwd("../../") # move back up to root folder
+        print(paste0("XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX      Results collation loop ", i, " complete        XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX"))
+  } # close i loop & go to the next i
 
   ####loops done create dfs####
   # create bin & Gaus barplot stats data frames
   if (bin) {binbars <- data.frame(Min.Inf = with(binbars.df, tapply(rel.inf, var, min)),
                                   Av.Inf = with(binbars.df, tapply(rel.inf, var, mean)),
                                   Max.Inf = with(binbars.df, tapply(rel.inf, var, max)),
-                                  Inf.variance = with(binbars.df, tapply(rel.inf, var, var)),
-                                  row.names = levels.default(binbars.df$var))
+                                  Inf.variance = with(binbars.df, tapply(rel.inf, var, var)))
   binbars <- binbars[order(-binbars[,"Av.Inf"]),]
   binbarsgood <- subset(binbars, Min.Inf > 0)} #also create barplots for nonzero vars
 
   if (gaus) {gausbars <- data.frame(Min.Inf = with(gausbars.df, tapply(rel.inf, var, min)),
                                     Av.Inf = with(gausbars.df, tapply(rel.inf, var, mean)),
                                     Max.Inf = with(gausbars.df, tapply(rel.inf, var, max)),
-                                    Inf.variance = with(gausbars.df, tapply(rel.inf, var, var)),
-                                    row.names = levels.default(gausbars.df$var))
+                                    Inf.variance = with(gausbars.df, tapply(rel.inf, var, var)))
   gausbars <- gausbars[order(-gausbars[,"Av.Inf"]),]
   gausbarsgood <- subset(gausbars, Min.Inf > 0)}
 
-  # create linesfiles end-column stats for each variable
+  # create linesfiles end-column stats (Min Av Max Var) for each variable
   if (bin) for (l in colnames(samples[expvar])) {
     assign(paste0("binline_", l), cbind(get(paste0("binline_", l)),
                                         "MinLine" = apply(get(paste0("binline_", l))[, (2:(1 + loops))], MARGIN = 1, min, na.rm = TRUE)))
@@ -358,42 +395,55 @@ gbm.loop <- function(loops = 10, # the number of loops required, integer
   if (bin) for (p in colnames(samples[expvar])) {
     yrange <- c(min(get(paste0("binline_", p))[,"MinLine"]), max(get(paste0("binline_", p))[,"MaxLine"]))
     png(filename = paste0("Bin_Loop_lines_", p, ".png"),
-        width = 4*480, height = 4*480, units = "px", pointsize = 80, bg = "white", res = NA, family = "", type = pngtype)
-    par(mar = c(2.3,5,0.3,0.4), fig = c(0,1,0,1), las = 1, lwd = 8, bty = "n", mgp = c(1.25,0.5,0), xpd = NA)
-    plot(get(paste0("binline_", p))[,1],
-         get(paste0("binline_", p))[,"AvLine"],
+        width = 4*480, height = 4*480, units = "px", pointsize = 40, bg = "white", res = NA, family = "", type = pngtype)
+    par(mar = c(2.3,5,0.3,0.4), fig = c(0,1,0,1), las = 1, lwd = 8, bty = "n", mgp = c(1.25, 0.5, 0), xpd = NA)
+    plot(if (class(get(paste0("binline_", p))[,1]) == "character") {x = 1:(length(get(paste0("binline_", p))[,1]))} else x = get(paste0("binline_", p))[,1],
+         # x doesn't work if it's a factor variable
+         y = get(paste0("binline_", p))[,"AvLine"],
          type = "l",
-         #xlab = colnames(get(paste0("binline_", p)))[1],
          xlab = paste0(p, " (", round(binbars[p, "Av.Inf"],1), "%)"),
          ylab = "",
          main = "",
-         yasx = "r",
-         ylim = yrange)
+         yaxs = "r",
+         ylim = yrange,
+         axes = FALSE)
+    # x axis labels
+    if (class(get(paste0("binline_", p))[,1]) == "character") {axis(1, at = 1:(length(get(paste0("binline_", p))[,1])), labels = get(paste0("binline_", p))[,1])} else axis(1)
+    axis(2) # y axis default
     mtext("Marginal Effect", side = 2, line = 4.05, las = 0)
-    lines(get(paste0("binline_", p))[,1], get(paste0("binline_", p))[,"MinLine"], col = "grey66") #[,1] is 1st column, X values, always the same
-    lines(get(paste0("binline_", p))[,1], get(paste0("binline_", p))[,"MaxLine"], col = "grey33")
-    legend("topleft", legend = c("Max","Av.","Min"), col = c("grey33","black","grey66"),
-           lty = 1, pch = "-")
+    lines(if (class(get(paste0("binline_", p))[,1]) == "character") {x = 1:(length(get(paste0("binline_", p))[,1]))} else x = get(paste0("binline_", p))[,1],
+          y = get(paste0("binline_", p))[,"MinLine"], col = "grey66") #[,1] is 1st column, X values, always the same
+    lines(if (class(get(paste0("binline_", p))[,1]) == "character") {x = 1:(length(get(paste0("binline_", p))[,1]))} else x = get(paste0("binline_", p))[,1],
+          y = get(paste0("binline_", p))[,"MaxLine"], col = "grey33")
+    legend("top", # Need to shrink & move this, but where/how to move? if() based on values?
+           legend = c("Max","Av.","Min"), col = c("grey33","black","grey66"),
+           lty = 1,
+           pch = "-")
     dev.off()}
 
   if (gaus) for (q in colnames(samples[expvar])) {
     yrange <- c(min(get(paste0("gausline_", q))[,"MinLine"]), max(get(paste0("gausline_", q))[,"MaxLine"]))
     png(filename = paste0("Gaus_Loop_lines_", q, ".png"),
-        width = 4*480, height = 4*480, units = "px", pointsize = 80, bg = "white", res = NA, family = "", type = pngtype)
+        width = 4*480, height = 4*480, units = "px", pointsize = 40, bg = "white", res = NA, family = "", type = pngtype)
     par(mar = c(2.3,5,0.3,0.4), fig = c(0,1,0,1), las = 1, lwd = 8, bty = "n", mgp = c(1.25,0.5,0), xpd = NA)
-    plot(get(paste0("gausline_", q))[,1],
-         get(paste0("gausline_", q))[,"AvLine"],
+    plot(if (class(get(paste0("gausline_", q))[,1]) == "character") {x = 1:(length(get(paste0("gausline_", q))[,1]))} else x = get(paste0("gausline_", q))[,1],
+         y = get(paste0("gausline_", q))[,"AvLine"],
          type = "l",
-         #xlab = colnames(get(paste0("gausline_", q)))[1],
          xlab = paste0(q, " (", round(gausbars[q, "Av.Inf"],1), "%)"),
          ylab = "",
          main = "",
-         yasx = "r",
-         ylim = yrange)
+         yaxs = "r",
+         ylim = yrange,
+         axes = FALSE)
+    # x axis labels
+    if (class(get(paste0("gausline_", q))[,1]) == "character") {axis(1, at = 1:(length(get(paste0("gausline_", q))[,1])), labels = get(paste0("gausline_", q))[,1])} else axis(1)
+    axis(2) # y axis default
     mtext("Marginal Effect", side = 2, line = 4.05, las = 0)
-    lines(get(paste0("gausline_", q))[,1], get(paste0("gausline_", q))[,"MinLine"], col = "grey66")
-    lines(get(paste0("gausline_", q))[,1], get(paste0("gausline_", q))[,"MaxLine"], col = "grey33")
-    legend("topleft", legend = c("Max","Av.","Min"), col = c("grey33","black","grey66"),
+    lines(if (class(get(paste0("gausline_", q))[,1]) == "character") {x = 1:(length(get(paste0("gausline_", q))[,1]))} else x = get(paste0("gausline_", q))[,1],
+          y = get(paste0("gausline_", q))[,"MinLine"], col = "grey66") #[,1] is 1st column, X values, always the same
+    lines(if (class(get(paste0("gausline_", q))[,1]) == "character") {x = 1:(length(get(paste0("gausline_", q))[,1]))} else x = get(paste0("gausline_", q))[,1],
+          y = get(paste0("gausline_", q))[,"MaxLine"], col = "grey33")
+    legend("top", legend = c("Max","Av.","Min"), col = c("grey33","black","grey66"),
            lty = 1, pch = "-")
     dev.off()}